#56 Sudoku games can now be saved and resumed later
- Completely revamped the beginning menu to a New/Resume screen
- Singleton game manager was created to easily sync game state between components
- Several components were simplified/updated to accommodate changes
