+++ src/app/@components/confirm/confirm.component.html
@@ -7,10 +7,10 @@
 
     <div id="confirm-actions" class="row g-0 align-items-center justify-content-end p-3">
         <div class="col-auto me-3">
-            <button (click)="confirm()">{{ confirmText }}</button>
+            <button class="confirm-button" (click)="confirm()">{{ confirmText }}</button>
         </div>
         <div class="col-auto">
-            <button (click)="cancel()">{{ cancelText }}</button>
+            <button class="cancel-button" (click)="cancel()">{{ cancelText }}</button>
         </div>
     </div>
 </div>
