+++ src/app/games/games-home.component.scss
@@ -25,8 +25,7 @@
     }
 
     .coming-soon {
-        background-color: rgba($color-slate, .5);
-        color: $color-gray;
+        color: $color-white;
         cursor: not-allowed;
         font-size: $default-font-size * 3;
         left: 0;
