+++ src/app/games/sudoku/help-dialog/help-dialog.component.html
@@ -11,14 +11,16 @@
     </div>
 
     <div class="row g-0 align-items-center p-3">
-        <p>Fill in the grid so that each row, column and 3x3 square contain one instance of the digits 1-9. There is a single solution, where each digit appears in the grid exactly 9 times.</p>
+        <p>Fill in the grid so that each row, column and 3x3 square contain one instance of the digits 1-9. There is a single solution, where each digit appears in the grid exactly nine times.</p>
         <ul class="px-3 m-0">
             <li><div>Enable the Pencil In toggle to help track candidate digits</div></li>
             <li><div>Click <i class="fas fa-cog"></i> to change game settings</div></li>
             <li><div>Click <i class="fas fa-ellipsis-h"></i> for more actions:</div>
                 <ul>
-                    <li><div>Check Cell -- reveal if the cell contains the correct digit</div></li>
+                    <li><div>Check Cell -- check if the cell contains the correct digit</div></li>
                     <li><div>Check All -- perform the Check Cell action for all cells</div></li>
+                    <li><div>Reveal Cell -- reveal the solution for one cell</div></li>
+                    <li><div>Reveal All -- reveal the entire solution (ends game)</div></li>
                     <li><div>Reset -- clear all cells and reset the game timer</div></li>
                 </ul>
             </li>
