+++ src/app/games/sudoku/menu-screen/menu-screen.models.ts
@@ -1,0 +1,4 @@
+export enum SudokuMenuState {
+    main = 1,
+    new = 2
+}
