+++ src/app/games/sudoku/settings-dialog/settings-dialog.component.html
@@ -10,41 +10,64 @@
         </div>
     </div>
 
-    <div *ngIf="hardcore" class="px-3">
-        <i class="fas fa-skull color-red"></i>
-        <span class="ms-3">Assistive settings are disabled in hardcore mode!</span>
+    <div class="setting-group">
+        <h5>General</h5>
+        <div class="row g-0 setting">
+            <div class="col">
+                <label class="d-block mt-2">Show timer</label>
+            </div>
+            <div class="col-auto ms-3">
+                <ak-toggle [(value)]="gameManager.gameSettings.showTimer" (valueChange)="toggleChange('showTimer', $event)"></ak-toggle>
+            </div>
+        </div>
+    </div>
+    
+    <div class="setting-group">
+        <h5>
+            <span>Assistive</span>
+            <ng-container *ngIf="gameManager.gameInstance?.hardcore">
+                <i class="fas fa-skull color-red ms-4 me-2"></i>
+                <span class="text-transform-none">Disabled in hardcore mode!</span>
+            </ng-container>
+        </h5>
+        <div class="row g-0 setting">
+            <div class="col">
+                <label class="d-block mt-2">Show conflicts</label>
+            </div>
+            <div class="col-auto ms-3">
+                <ak-toggle [disabled]="gameManager.gameInstance?.hardcore" [(value)]="gameManager.gameSettings.showConflicts" (valueChange)="toggleChange('showConflicts', $event)"></ak-toggle>
+            </div>
+        </div>
+        <div class="row g-0 setting">
+            <div class="col">
+                <label class="d-block mt-2">Erase pencil</label>
+                <p class="setting-description">After entering a number, any conflicting penciled in values in the same row, column and 3x3 square will be erased</p>
+            </div>
+            <div class="col-auto ms-3">
+                <ak-toggle [disabled]="gameManager.gameInstance?.hardcore" [(value)]="gameManager.gameSettings.erasePencil" (valueChange)="toggleChange('erasePencil', $event)"></ak-toggle>
+            </div>
+        </div>
+        <div class="row g-0 setting">
+            <div class="col">
+                <label class="d-block mt-2">Disable inputs</label>
+                <p class="setting-description">Number inputs will be disabled when the count for their values reaches nine</p>
+            </div>
+            <div class="col-auto ms-3">
+                <ak-toggle [disabled]="gameManager.gameInstance?.hardcore" [(value)]="gameManager.gameSettings.disableInputs" (valueChange)="toggleChange('disableInputs', $event)"></ak-toggle>
+            </div>
+        </div>
     </div>
 
-    <div class="row g-0 align-items-center setting">
-        <div class="col">
-            <label>Show clock</label>
-        </div>
-        <div class="col-auto">
-            <ak-toggle [(value)]="showClock" (valueChange)="toggleChange('sudoku_showClock', $event)"></ak-toggle>
-        </div>
-    </div>
-    <div class="row g-0 align-items-center setting">
-        <div class="col">
-            <label>Show conflicts</label>
-        </div>
-        <div class="col-auto">
-            <ak-toggle [disabled]="hardcore" [(value)]="showConflicts" (valueChange)="toggleChange('sudoku_showConflicts', $event)"></ak-toggle>
-        </div>
-    </div>
-    <div class="row g-0 align-items-center setting">
-        <div class="col">
-            <label>Auto pencil erase</label>
-        </div>
-        <div class="col-auto">
-            <ak-toggle [(value)]="autoPencilErase" (valueChange)="toggleChange('sudoku_autoPencilErase', $event)"></ak-toggle>
-        </div>
-    </div>
-    <div class="row g-0 align-items-center setting">
-        <div class="col">
-            <label>Auto disable inputs</label>
-        </div>
-        <div class="col-auto">
-            <ak-toggle [disabled]="hardcore" [(value)]="autoDisableInputs" (valueChange)="toggleChange('sudoku_autoDisableInputs', $event)"></ak-toggle>
+    <div class="setting-group">
+        <h5>Game Saves</h5>
+        <div class="row g-0 setting">
+            <div class="col">
+                <label class="d-block mt-2">Exit Save</label>
+                <p class="setting-description">Sudoku progress will be automatically saved when exiting the game</p>
+            </div>
+            <div class="col-auto ms-3">
+                <ak-toggle [(value)]="gameManager.gameSettings.exitSave" (valueChange)="toggleChange('exitSave', $event)"></ak-toggle>
+            </div>
         </div>
     </div>
 </div>
