+++ src/app/games/sudoku/solved-dialog/solved-dialog.component.ts
@@ -1,11 +1,13 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { Component, OnInit, inject } from '@angular/core';
 import { DialogBase } from '@directives/dialog/dialog-base';
 import { DifficultyPipe } from '@pipes/difficulty.pipe';
-import { SudokuDifficulty } from '../sudoku.models';
+import { SudokuManager } from '../sudoku-manager';
 
 @Component({
     standalone: true,
     imports: [
+        CommonModule,
         DifficultyPipe
     ],
     selector: 'ak-solved-dialog',
@@ -13,15 +15,14 @@ import { SudokuDifficulty } from '../sudoku.models';
     templateUrl: './solved-dialog.component.html'
 })
 export class SolvedDialogComponent extends DialogBase implements OnInit {
-    @Input() difficulty: SudokuDifficulty;
-    @Input() hardcore: boolean;
-    @Input() time: string;
-    @Input() goHome: () => void;
-    @Input() playAgain: () => void;
+    public gameManager = inject(SudokuManager);
+
+    public goHome: () => void;
+    public playAgain: () => void;
 
     public timeDisplay = '';
 
     public ngOnInit(): void {
-        this.timeDisplay = this.time.trimLeftChars(['0', ':']);
+        this.timeDisplay = this.gameManager.gameInstance.time$.value.trimLeftChars(['0', ':']);
     }
 }
