+++ src/style/ak-app.scss
@@ -57,25 +57,7 @@ section {
     }
 }
 
-h1 {
-    font-weight: bold;
-    margin: 0;
-    text-transform: uppercase;
-}
-
-h2 {
-    font-weight: bold;
-    margin: 0;
-    text-transform: uppercase;
-}
-
-h3 {
-    font-weight: bold;
-    margin: 0;
-    text-transform: uppercase;
-}
-
-h4 {
+h1, h2, h3, h4, h5 {
     font-weight: bold;
     margin: 0;
     text-transform: uppercase;
@@ -120,6 +102,10 @@ button {
     font-family: $default-font-family;
     padding: math.div($base-size, 2);
 
+    &:disabled {
+        opacity: .5;
+    }
+
     &:not(:disabled) {
         cursor: pointer;
 
@@ -138,6 +124,22 @@ button {
         &:not(:disabled):hover {
             background-color: transparent;
             color: $color-blue;
+        }
+    }
+
+    &.confirm-button {
+        &:hover, &:focus {
+            background-color: $color-green;
+            border-color: $color-green;
+            color: $color-white;
+        }
+    }
+
+    &.cancel-button {
+        &:hover, &:focus {
+            background-color: $color-red;
+            border-color: $color-red;
+            color: $color-white;
         }
     }
 }
