+++ src/style/ak-helpers.scss
@@ -70,3 +70,8 @@
     fill: $color-error;
     stroke: $color-error;
 }
+
+// Text/font classes
+.text-transform-none {
+    text-transform: none;
+}
